+++ test/deployment/BaoFactory.t.sol
@@ -5,7 +5,9 @@ import {Test, Vm} from "forge-std/Test.sol";
 import {LibClone} from "@solady/utils/LibClone.sol";
 import {UUPSUpgradeable} from "@solady/utils/UUPSUpgradeable.sol";
 
-import {BaoFactory, BaoFactoryLib, IBaoFactory} from "@bao-script/deployment/BaoFactory.sol";
+import {BaoFactory} from "@bao/factory/BaoFactory.sol";
+import {BaoFactoryLib} from "@bao/factory/BaoFactoryLib.sol";
+import {IBaoFactory} from "@bao-factory/IBaoFactory.sol";
 import {FundedVault, NonPayableVault, FundedVaultUUPS} from "@bao-test/mocks/deployment/FundedVault.sol";
 import {BaoFactoryV2, BaoFactoryNonUUPS} from "@bao-test/mocks/deployment/BaoFactoryV2.sol";
 
@@ -147,7 +149,7 @@ contract BaoFactoryTest is Test {
         vm.prank(owner);
         factory.setOperator(op2, 2 days);
 
-        (address[] memory addrs, uint40[] memory expiries) = factory.operators();
+        (address[] memory addrs, uint256[] memory expiries) = factory.operators();
 
         assertEq(addrs.length, 2);
         assertEq(expiries.length, 2);
